Dev v2.1.0

+++ src/hexutils.cpp
@@ -4,7 +4,7 @@ String toHex(uint8_t* in) {
 	return toHex(in, sizeof(in)*2);
 }
 
-String toHex(uint8_t* in, uint8_t size) {
+String toHex(uint8_t* in, uint16_t size) {
 	String hex;
 	for(int i = 0; i < size; i++) {
 		if(in[i] < 0x10) {
@@ -16,7 +16,7 @@ String toHex(uint8_t* in, uint8_t size) {
 	return hex;
 }
 
-void fromHex(uint8_t *out, String in, uint8_t size) {
+void fromHex(uint8_t *out, String in, uint16_t size) {
 	for(int i = 0; i < size*2; i += 2) {
 		out[i/2] = strtol(in.substring(i, i+2).c_str(), 0, 16);
 	}
